fix: profiler for VR
Note, AE appears to be broken, but I don't think it actually ever worked
with `Profiler-report = true`. The functions do not match SE addresses
at least with AE 1130. I have made some slight modifications to AE
addresses, but those were for AE1130 and should be checked with 1170.
The CTD on startup appears to be tied to `ConsoleOpen`'s hooked
function `func(playerCharacter*)` that was inlined by AE1130.